jms: cleans up artemis code in attempts to deflake test
