+++ instrumentation/jms-jakarta/src/it/jms30/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    Copyright 2013-2023 The OpenZipkin Authors
+    Copyright 2013-2024 The OpenZipkin Authors
 
     Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
     in compliance with the License. You may obtain a copy of the License at
@@ -20,13 +20,14 @@
   <modelVersion>4.0.0</modelVersion>
 
   <groupId>@project.groupId@</groupId>
-  <artifactId>jms40</artifactId>
+  <artifactId>jms30</artifactId>
   <version>@project.version@</version>
-  <name>jms40</name>
+  <name>jms30</name>
 
   <properties>
     <maven.compiler.source>1.8</maven.compiler.source>
     <maven.compiler.target>1.8</maven.compiler.target>
+    <maven.compiler.release>8</maven.compiler.release>
   </properties>
 
   <dependencies>
