+++ instrumentation/jms-jakarta/src/test/java/brave/jakarta/jms/ITTracingJMSConsumer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2023 The OpenZipkin Authors
+ * Copyright 2013-2024 The OpenZipkin Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
@@ -40,7 +40,7 @@ import static org.assertj.core.api.Assertions.assertThat;
  * {@link brave.jakarta.jms.ITTracingMessageConsumer}
  */
 class ITTracingJMSConsumer extends ITJms {
-  @RegisterExtension ArtemisJmsExtension jms = new ArtemisJmsExtension();
+  @RegisterExtension JmsExtension jms = new JmsExtension();
 
   JMSContext tracedContext;
   JMSProducer producer;
